健康数据相关接口

+++ src/main/java/com/scau/mapper/ExerciseRecordsMapper.java
@@ -1,7 +1,6 @@
 package com.scau.mapper;
 
 import com.scau.entity.healthData.pojo.ExerciseRecords;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Delete;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Param;
@@ -20,9 +19,9 @@ public interface ExerciseRecordsMapper {
 
 
     @Insert("INSERT INTO exercise_records values (0,#{userId},#{exerciseType},#{duration},#{distance},#{calories},#{heartRate},#{note},now())")
-    void insert(Long userId, String exerciseType, Integer duration, Double distance, Double calories, Integer heartRate, String note);
+    void insert(Long userId, String exerciseType, Double duration, Double distance, Double calories, Double heartRate, String note);
 
-    void updateById(Long id, Long userId, String exerciseType, Integer duration, Double distance, Double calories, Integer heartRate, String note);
+    void updateById(Long id, Long userId, String exerciseType, Double duration, Double distance, Double calories, Double heartRate, String note);
 
     @Delete("delete from exercise_records where exercise_id=#{id} and user_id=#{userId}")
     void deleteById(Long id, Long userId);
