+++ src/main/java/com/scau/mapper/VitalSignsMapper.java
@@ -5,7 +5,6 @@ import org.apache.ibatis.annotations.Delete;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Param;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -19,9 +18,9 @@ public interface VitalSignsMapper {
     List<VitalSigns> selectList(@Param("userId")Long userId, @Param("startIndex") int startIndex, @Param("limit") Integer limit, @Param("keyword") String keyword);
 
     @Insert("INSERT INTO vital_signs values (0,#{userId},#{height},#{weight},#{bloodPressure},#{bloodSugar},#{bloodLipids},#{note},now())")
-    void insert(@Param("userId") Long userId, @Param("height")Double height, @Param("weight") Double weight, @Param("bloodPressure") Integer bloodPressure, @Param("bloodSugar") Double bloodSugar, @Param("bloodLipids")Double bloodLipids, @Param("note")String note);
+    void insert(@Param("userId") Long userId, @Param("height")Double height, @Param("weight") Double weight, @Param("bloodPressure") Double bloodPressure, @Param("bloodSugar") Double bloodSugar, @Param("bloodLipids")Double bloodLipids, @Param("note")String note);
 
-    void updateById(Long id, Long userId, Double height, Double weight, Integer bloodPressure, Double bloodSugar, Double bloodLipids, String note);
+    void updateById(Long id, Long userId, Double height, Double weight, Double bloodPressure, Double bloodSugar, Double bloodLipids, String note);
 
     @Delete("delete from vital_signs where record_id=#{id} and user_id=#{userId}")
     void deleteById(Long id, Long userId);
